requirements.txt
flask==3.0.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
flask-login==0.6.3
Werkzeug==3.0.3
requests==2.32.5

